Rename unit test package from unit to ut to match existing and integration test package naming
