+++ imi/src/test/java/org/motechproject/nms/imi/ut/FileProcessedStatusTest.java
@@ -1,11 +1,11 @@
-package org.motechproject.nms.imi.unit;
+package org.motechproject.nms.imi.ut;
 
 import org.junit.Test;
 import org.motechproject.nms.imi.domain.FileProcessedStatus;
 
 import static junit.framework.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 public class FileProcessedStatusTest {
     @Test
