+++ imi/src/test/java/org/motechproject/nms/imi/ut/RequestIdTest.java
@@ -1,4 +1,4 @@
-package org.motechproject.nms.imi.unit;
+package org.motechproject.nms.imi.ut;
 
 import org.junit.Test;
 import org.motechproject.nms.imi.service.RequestId;
